Add files via upload
CharacterProperty.java is the class that defines what a character property is (font, size, color)

CharacterPropertyFactory.java is the factory that takes font, size, color, then checks to see if the same attributes are all present within an already created property. If so, return the already created property. If not, create a new property and save it.

Document.java is a class that defines what a Document object is. String content, String name, and Character Property

TextEditor.java is the main driver program

FlyweightTest.java contains the unit tests