Add files via upload
Song.java is the class that defines what a Song object should be: title, artist, album, duration

SongService.java is the interface provided in the assignment. Defines the functions of a song service

SongServer.java implements SongService and is the main server for the music service. It stores the Song objects for all songs

SongProxy.java implements SongService and is the proxy to the main server for the music service. It stores Song objects for some songs, fetching from the main server when needed

MusicStreamingDriver.java is the main driver program

ProxyTest.java contains the unit tests